+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Show project ID after project creation
 
+### Fixed
+
+- Aliased dependency names in `package-lock.json`
+
 ## [5.8.1] - 2023-11-07
 
 ### Fixed